refactor the marker hooks and services
