+++ src/components/Map/hooks/useMarkers.js
@@ -1,54 +1,15 @@
 /* eslint-disable no-loop-func */
 /* eslint-disable no-unused-vars */
-import { useRef, useEffect, useState } from 'react';
+import { useRef, useEffect, useState, useMemo } from 'react';
 import L from 'leaflet';
 import 'leaflet.markercluster';
 
-import { categories, CATEGORY_NAMES } from '../static/categories';
+import { categories} from 'static/categories';
+import { removeCurrentHighlight, HIGHLIGHT_CLASS , getMarkerIcon } from 'services/marker';
+import useHighlight from './useHighlight';
 
-const markerOptions = {
-  iconSize: [23, 23],
-  iconAnchor: [8, 15],
-  popupAnchor: [-3, -76],
-};
 
-export const HIGHLIGHT_CLASS = 'active-element';
-
-let markerHighlight;
 let areaHighlightLayer;
-
-export function getMarkerCategory(device) {
-  return categories[Object.keys(categories).find(mt => mt === device.categories[0])];
-}
-
-function getMarkerIcon(categoryName) {
-  const name = categoryName === 'Beacons' ? 'Baken' : categoryName;
-  const iconUrl = categories[name].iconUrl;
-  return L.icon({
-    ...markerOptions,
-    iconUrl,
-  });
-}
-
-export const createFeatureMarker = (latlng, category) =>
-  // Create a marker with the correct icon and onClick method
-  L.marker(latlng, {
-    icon: getMarkerIcon(category),
-  });
-
-export const removeCurrentHighlight = map => {
-  if (markerHighlight) {
-    // Point highlight
-    map.removeLayer(markerHighlight);
-  }
-
-  if (areaHighlightLayer) {
-    const classList = areaHighlightLayer.getElement().classList;
-    if (classList && classList.remove) {
-      classList.remove(HIGHLIGHT_CLASS);
-    }
-  }
-};
 
 export const showDeviceInfo = (event, marker, onClick, highlight) => {
   highlight(event);
@@ -67,35 +28,19 @@ export const showAreaInfo = (event, map, onClick) => {
   onClick(event.sourceTarget.feature);
 };
 
-const clusterCategories = Object.entries(categories).filter(([, value]) => value.isClustered);
 
-const useHighlight = () => {
-  const [activeMarker, setActiveMarker] = useState();
-
-  const highlight = ({ sourceTarget }) => {
-    if (activeMarker) {
-      activeMarker._icon.classList.remove('highlight');
-    }
-
-    setActiveMarker(sourceTarget);
-  };
-
-  useEffect(() => {
-    if (activeMarker) activeMarker._icon.classList.add('highlight');
-  }, [activeMarker]);
-
-  return highlight;
-};
-
-export const useMarkers = map => {
+const useMarkers = map => {
   const markerGroupRef = useRef(null);
   const layerListRef = useRef({});
   const layerGroupRef = useRef({});
   const highlight = useHighlight();
 
+  const clusterCategories = useMemo(() => Object.entries(categories).filter(([, value]) => value.isClustered), []);
+
+
   useEffect(() => {
     markerGroupRef.current = L.markerClusterGroup({
-      disableClusteringAtZoom: 8,
+      disableClusteringAtZoom: 16,
       showCoverageOnHover: false,
       spiderfyOnMaxZoom: false,
     });
@@ -159,24 +104,7 @@ export const useMarkers = map => {
       });
   };
 
-  const addPrivacyLayers = (layers, onClickCallback) => {
-    layers.forEach(layerData => {
-      const { name, layer, category } = layerData;
-
-      if (!layerListRef.current[name]) {
-        layerListRef.current[name] = L.Proj.geoJson(layer, {
-          pointToLayer: (feature, latlng) => createFeatureMarker(latlng, category),
-        });
-        if (map) map.addLayer(layerListRef.current[name]);
-      }
-
-      const mapLayer = layerListRef.current[name];
-      mapLayer.addData(layer);
-      mapLayer.on('click', event => showAreaInfo(event, map, onClickCallback));
-      layerGroupRef.current[category] = layerGroupRef.current[category] || [];
-      layerGroupRef.current[category].push(name);
-    });
-  };
-
-  return { addMarkers, addAreas, toggleLayer, addPrivacyLayers };
+  return { addMarkers, addAreas, toggleLayer};
 };
+
+export default useMarkers;
