+++ src/services/api/privacyLayersConfig.js
@@ -46,6 +46,7 @@ const PRIVACY_LAYERS_CONFIG = [
       ...item,
       category: CATEGORY_NAMES.SLIMME_VERKEERSINFORMATIE,
       soort: item.properties.Locatienaam,
+      privacy: '',
       contact: 'ais-masten',
       longitude: item.geometry.coordinates[0],
       latitude: item.geometry.coordinates[1],
@@ -75,6 +76,7 @@ const PRIVACY_LAYERS_CONFIG = [
       ...item,
       category: CATEGORY_NAMES.CAMERA,
       soort: item.properties.Soort,
+      privacy: '',
       contact: 'overig',
       longitude: item.geometry.coordinates[0],
       latitude: item.geometry.coordinates[1],
